move solution to seperate folder, make master post route work
